Moved test to BulkFinalizationTest
